clean up basic rearrangement moves and add tests
