+++ src/phylox/exceptions/invalidmoveexception.py
@@ -1,0 +1,7 @@
+class InvalidMoveException(Exception):
+    """Exception raised when trying to apply an invalid move.
+
+    Attributes:
+        message -- explanation of the error
+    """
+    pass
