+++ src/phylox/isomorphism/__init__.py
@@ -1,0 +1,1 @@
+from .base import is_isomorphic
