+++ src/phylox/rearrangement/heuristics/red_line_heuristic.py
